+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Add ArrayTypeUtils with operations on array types
 * Add TypeVariableUtils for working with type templates (with preserved variables)
 * Add TypeLiteral utility class to simplify complex types declaration (by analogy with guice's class)
 * GenericUtils